Uniffi

+++ src/errors/mod.rs
@@ -3,13 +3,14 @@ use binary_codec::SerializationError as BinarySerializationError;
 use serde::Serialize;
 
 /// Plabble Serialization Error (based on binary_codec errors)
+#[cfg_attr(feature = "ffi", derive(uniffi::Error))]
 #[derive(Debug, PartialEq, Serialize)]
 pub enum SerializationError {
     /// Value is out of bounds (value, min, max)
     ValueOutOfBounds(i32, i32, i32),
 
     // Unexpected size (expected, actual)
-    UnexpectedLength(usize, usize),
+    UnexpectedLength(u64, u64),
 
     /// Missing runtime length key
     MissingLengthByKey(String),
@@ -42,13 +43,14 @@ impl From<BinarySerializationError> for SerializationError {
 }
 
 /// Plabble Deserialization Error (based on binary_codec errors)
+#[cfg_attr(feature = "ffi", derive(uniffi::Error))]
 #[derive(Debug, PartialEq, Serialize)]
 pub enum DeserializationError {
     /// Not enough bytes (bytes missing)
-    NotEnoughBytes(usize),
+    NotEnoughBytes(u64),
 
     // Unexpected size (expected, actual)
-    UnexpectedLength(usize, usize),
+    UnexpectedLength(u64, u64),
 
     /// Unknown enum discriminator
     UnknownDiscriminant(u8),
